requirements.txt
google-genai
langgraph
aiohttp
pydantic